refactor(#3303): split tui_direct_abort_marker into mod.rs + store.rs (pure move)
Directory-decompose the marker module ahead of the deferred-claim marker
work: the durable store + commit-tombstone I/O (root/record/load/delete/
reload/tombstone files/GC) and the sidecar-flock claim move to
tui_direct_abort_marker/store.rs (348 prod lines); mod.rs (691 prod lines,
was 999/1000) keeps the marker types, pure decision functions, reaction
applier, and the drain/sweep reconcilers. Every external caller path is
preserved via pub(in crate::services::discord) re-exports — tmux_watcher /
placeholder_sweeper / tui_prompt_relay are 0-line changes, and all 30
existing marker tests pass unmodified.
